# Improve throughput on Compy