Improve throughput on Compy by reducing stdout writes:  Turn off MOSART debug writes and increase tolerance in an ELM BalanceCheck.  Also add a throughput tolerance check to compy.

Fixes E3SM-Project/E3SM#3150

[BFB]