@amametjanov 2019-10-14
Before these updates:
```
    TOT Run Time:     821.140 seconds       26.488 seconds/mday         8.94 myears/wday
```
After:
```
    TOT Run Time:     668.517 seconds       21.565 seconds/mday        10.98 myears/wday
```
The new throughput (with today's master) is now within 2% (as opposed to 20% slower) compared to throughput of the same case built with master from July 30 (hash `a089690`):
```
    TOT Run Time:     656.696 seconds       21.184 seconds/mday        11.17 myears/wday
```