@singhbalwinder 2019-10-15
So, what are the reasons for this performance improvement? Is it simply the removal of log file write statements or are there some other flags involved as well?