@rljacob 2019-10-15
Writing lots of output to stdout every timestep is expensive.